Added functionality and UI
Created a UI for a pause menu, a game over panel and the in game UI.
Added player movement functionality with a touch "analogue stick". Speed needs to be adjusted.
Added a jump button, currently bugged, can only jump once or not at all.
Added a ball and hoops. Score is counted. Play area needs to be built.