Made a number of changes
Modified the arena and a number of sizes of objects within the scene.
High score is displayed on the title screen, and is updated when beaten.
The hoops can provide 20 points, if the ball goes through without hitting the ring, and 10 points if it hits the rim. Doesn't give points if the ball goes upwards through the hoop.
Player has extra collision detection, to prevent walking through walls.

Countdown timer currently doesn't work, needs to be modified for this updated version of the game.