Many tweaks and changes
Removed cool down timer after scoring a single point.
Added a height requirement for the ball after its collided with the player. If it doesn't reach that height before touching the player again, no point is awarded. This is to prevent holding the ball on the players head.
Joystick now works correctly over varying screen sizes.
Slightly adjusted the size of the ball and hoops.