Google achievements and leaderboard functional
You can now unlock achievements and post score to the google leaderboards.
Google achievements don't pop up when they have be achieved. I think this may be due to the game running in landscape mode. The achievements are collected however.

Added icons to to achievement and leaderboard button.

Added a counter for the number of hoops and swishes scored in a game. These values are stored in their own google leaderboard. The swish and hoop scores are visible when the game has ended.

Fixed floating ball problem when game hasn't begun