Made a few changes based on user feedback
Added the music sound and control toggles to the title page, shown by clicking the new button. Currently doesn't have an icon.
Settings panel slides up and down when button is clicked.
Changed the camera to have a solid colour instead of a skybox. Potential fix for poor performance.
Added a return to main button into the pause panel
Pause button moved to in front of the left jump button.
Modified the jump button and made the joystick smaller

Added the firebase analytics package for use in the future