Lots of changes made
Ads have been implemented into the game, both rewarded and video ads. Video ads rarely play, and don't at all if rewarded ads are being used.

Changed the shaders to more mobile friendly shaders, massively increasing performace.

Tips on how to play are displayed on the screen when the game loads. Fades when the countdown starts.

Players velocity is zeroed when the position is reset on ball drop.

The music now loops as long as the game is running.

A confetti explosion appears from the hoop when the player scores a hoop or a swish, accompanied by text flying at the screen.

Changed SpherePhysics to BallController.
Changed ColliderControl to BallGoalColliderControl.